requirements.txt
flask
tldextract
dnspython
